refactor: consolidate sync utilities and remove redundant code
