+++ src/db/sync-service.ts
@@ -6,26 +6,16 @@
 import type { CRSQLChange } from './crsqlite-worker'
 import { getLatestMigrationVersion } from './migrate'
 import { DatabaseSingleton } from './singleton'
-
-const SYNC_SERVER_VERSION_KEY = 'thunderbolt_server_version'
-const SITE_ID_KEY = 'thunderbolt_site_id'
-const SYNC_VERSION_KEY = 'thunderbolt_sync_version'
-
-/**
- * Serialized change format for network transport
- * Uses base64 for binary data (pk, site_id)
- */
-export type SerializedChange = {
-  table: string
-  pk: string // base64 encoded
-  cid: string
-  val: unknown
-  col_version: string // bigint as string
-  db_version: string // bigint as string
-  site_id: string // base64 encoded
-  cl: number
-  seq: number
-}
+import {
+  type SerializedChange,
+  deserializeChange,
+  getLastSyncedVersion,
+  getServerVersion,
+  getSiteId,
+  serializeChange,
+  setLastSyncedVersion,
+  setServerVersion,
+} from './sync-utils'
 
 /**
  * WebSocket message types
@@ -57,56 +47,6 @@ export type SyncServiceOptions = {
   /** Called when chat sessions have been updated from remote changes */
   onChatSessionsChanged?: (chatThreadIds: string[]) => void
 }
-
-/**
- * Encode Uint8Array to base64 string
- */
-const encodeBase64 = (data: Uint8Array): string => {
-  const bytes = Array.from(data)
-  return btoa(String.fromCharCode(...bytes))
-}
-
-/**
- * Decode base64 string to Uint8Array
- */
-const decodeBase64 = (base64: string): Uint8Array => {
-  const binary = atob(base64)
-  const bytes = new Uint8Array(binary.length)
-  for (let i = 0; i < binary.length; i++) {
-    bytes[i] = binary.charCodeAt(i)
-  }
-  return bytes
-}
-
-/**
- * Serialize a CRSQLChange for network transport
- */
-const serializeChange = (change: CRSQLChange): SerializedChange => ({
-  table: change.table,
-  pk: encodeBase64(change.pk),
-  cid: change.cid,
-  val: change.val,
-  col_version: change.col_version.toString(),
-  db_version: change.db_version.toString(),
-  site_id: encodeBase64(change.site_id),
-  cl: change.cl,
-  seq: change.seq,
-})
-
-/**
- * Deserialize a network change to CRSQLChange
- */
-const deserializeChange = (serialized: SerializedChange): CRSQLChange => ({
-  table: serialized.table,
-  pk: decodeBase64(serialized.pk),
-  cid: serialized.cid,
-  val: serialized.val,
-  col_version: BigInt(serialized.col_version),
-  db_version: BigInt(serialized.db_version),
-  site_id: decodeBase64(serialized.site_id),
-  cl: serialized.cl,
-  seq: serialized.seq,
-})
 
 export class SyncService {
   private wsUrl: string
@@ -169,36 +109,6 @@ export class SyncService {
     }
   }
 
-  private getServerVersion(): bigint {
-    const stored = localStorage.getItem(SYNC_SERVER_VERSION_KEY)
-    return stored ? BigInt(stored) : 0n
-  }
-
-  private setServerVersion(version: bigint): void {
-    localStorage.setItem(SYNC_SERVER_VERSION_KEY, version.toString())
-  }
-
-  private getLastSyncedVersion(): bigint {
-    const stored = localStorage.getItem(SYNC_VERSION_KEY)
-    return stored ? BigInt(stored) : 0n
-  }
-
-  private setLastSyncedVersion(version: bigint): void {
-    localStorage.setItem(SYNC_VERSION_KEY, version.toString())
-  }
-
-  async getSiteId(): Promise<string> {
-    const storedSiteId = localStorage.getItem(SITE_ID_KEY)
-    if (storedSiteId) {
-      return storedSiteId
-    }
-
-    const db = DatabaseSingleton.instance.syncableDatabase
-    const siteId = await db.getSiteId()
-    localStorage.setItem(SITE_ID_KEY, siteId)
-    return siteId
-  }
-
   private async connect(): Promise<void> {
     if (!DatabaseSingleton.instance.supportsSyncing) {
       return
@@ -223,7 +133,7 @@ export class SyncService {
         this.reconnectAttempts = 0
 
         // Authenticate
-        const siteId = await this.getSiteId()
+        const siteId = await getSiteId()
         const migrationVersion = getLatestMigrationVersion()
 
         this.send({
@@ -269,10 +179,10 @@ export class SyncService {
       case 'auth_success': {
         console.info('WebSocket authenticated, server version:', response.serverVersion)
         this.setStatus('connected')
-        this.setServerVersion(BigInt(response.serverVersion))
+        setServerVersion(BigInt(response.serverVersion))
 
         // Request any changes we might have missed while offline
-        const serverVersion = this.getServerVersion()
+        const serverVersion = getServerVersion()
         this.send({ type: 'pull', since: serverVersion.toString() })
 
         // Set up database change listener for real-time push
@@ -293,8 +203,8 @@ export class SyncService {
 
       case 'push_success': {
         console.info('Push successful, server version:', response.serverVersion)
-        this.setServerVersion(BigInt(response.serverVersion))
-        this.setLastSyncedVersion(this.lastPushedVersion)
+        setServerVersion(BigInt(response.serverVersion))
+        setLastSyncedVersion(this.lastPushedVersion)
         this.isPushingChanges = false
 
         // Push any pending changes that accumulated during the push
@@ -339,7 +249,7 @@ export class SyncService {
           }
         }
 
-        this.setServerVersion(BigInt(response.serverVersion))
+        setServerVersion(BigInt(response.serverVersion))
         break
       }
 
@@ -439,7 +349,7 @@ export class SyncService {
 
     try {
       const db = DatabaseSingleton.instance.syncableDatabase
-      const lastSyncedVersion = this.getLastSyncedVersion()
+      const lastSyncedVersion = getLastSyncedVersion()
       const { changes, dbVersion } = await db.getChanges(lastSyncedVersion)
 
       if (changes.length === 0) {
@@ -537,7 +447,7 @@ export class SyncService {
 
     await this.pushLocalChanges()
 
-    const serverVersion = this.getServerVersion()
+    const serverVersion = getServerVersion()
     this.send({ type: 'pull', since: serverVersion.toString() })
   }
 }
